fix(border): helper function to apply border style not working as it should
