+++ packages/admin-ui-core/src/styles.ts
@@ -8,8 +8,7 @@ import { callUtil, isUtil } from './utils'
 import { createTransform } from './transforms'
 import type { StyleObject, StyleProp } from './types'
 import { theme as defaultTheme } from './theme'
-
-const TOKEN_PREFIX = '$'
+import { extractTokenCall } from './helpers'
 
 /**
  * Parses a style object
@@ -55,13 +54,3 @@ export function createCsx(theme?: any) {
 
   return csx
 }
-
-export function isToken(token: string) {
-  return typeof token === 'string' && token.startsWith(TOKEN_PREFIX)
-}
-
-export function extractTokenCall(token: string) {
-  return isToken(token) ? token.substring(1) : token
-}
-
-export const cx = (...args: string[]) => args.join(' ')
